requirements.txt
django==6.0.3
djangorestframework==3.17.0
djangorestframework-xml==2.0.0
defusedxml==0.7.1
psycopg2-binary==2.9.11
split-settings==1.0.0
djangorestframework_simplejwt==5.5.1